# Tree: add location for most nodes