This closes #732 (although some node types do not have locations yet e.g. array and struct initializers)

As expected this increases memory usage but there's not much of a performance hit (numbers are for parsing `zig2.c` which is `161,279,661` bytes)

```
Before
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          9.54s  ± 9.92ms    9.53s  … 9.55s           0 ( 0%)        0%
  peak_rss           1.61GB ± 28.8KB    1.61GB … 1.61GB          0 ( 0%)        0%
  cpu_cycles         36.9G  ± 40.2M     36.8G  … 36.9G           0 ( 0%)        0%
  instructions       78.6G  ± 70.2      78.6G  … 78.6G           0 ( 0%)        0%
  cache_references   93.3M  ±  399K     92.9M  … 93.7M           0 ( 0%)        0%
  cache_misses       43.9M  ±  187K     43.6M  … 44.0M           0 ( 0%)        0%
  branch_misses       327M  ±  345K      327M  …  327M           0 ( 0%)        0%

After
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          9.63s  ± 40.2ms    9.60s  … 9.67s           0 ( 0%)          +  1.0% ±  0.7%
  peak_rss           1.74GB ± 59.1KB    1.74GB … 1.74GB          0 ( 0%)        💩+  7.8% ±  0.0%
  cpu_cycles         37.0G  ±  167M     36.9G  … 37.2G           0 ( 0%)          +  0.5% ±  0.7%
  instructions       78.7G  ±  101      78.7G  … 78.7G           0 ( 0%)          +  0.2% ±  0.0%
  cache_references   97.1M  ±  808K     96.4M  … 98.0M           0 ( 0%)        💩+  4.1% ±  1.5%
  cache_misses       46.7M  ±  328K     46.4M  … 47.1M           0 ( 0%)        💩+  6.4% ±  1.4%
  branch_misses       327M  ±  599K      326M  …  327M           0 ( 0%)          -  0.1% ±  0.3%
```