@ehaas 2024-08-15
Looks like we're massively faster than GCC with similar memory usage, and about 40% faster than clang at the cost of 44% higher memory usage. The only caveat is that I'm not sure if those 2 are doing more work than we are for `-fsyntax-only`


```
Benchmark 1 (3 runs): clang-18 -fsyntax-only -I/home/ehaas/source/zig/lib /home/ehaas/source/zig/build/zig2.c
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          16.2s  ± 41.8ms    16.2s  … 16.3s           0 ( 0%)        0%
  peak_rss           1.20GB ± 81.3KB    1.20GB … 1.20GB          0 ( 0%)        0%
  cpu_cycles         66.5G  ±  174M     66.3G  … 66.6G           0 ( 0%)        0%
  instructions       99.9G  ±  177M     99.8G  …  100G           0 ( 0%)        0%
  cache_references    418M  ± 8.74M      408M  …  424M           0 ( 0%)        0%
  cache_misses       36.9M  ±  188K     36.7M  … 37.1M           0 ( 0%)        0%
  branch_misses       126M  ±  454K      126M  …  127M           0 ( 0%)        0%
Benchmark 2 (3 runs): /home/ehaas/source/arocc/zig-out/bin/arocc-node-locs -fsyntax-only -I/usr/lib/gcc/x86_64-linux-gnu/10/include -I/home/ehaas/source/zig/lib /home/ehaas/source/zig/build/zig2.c
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          9.58s  ± 38.5ms    9.55s  … 9.62s           0 ( 0%)        ⚡- 40.9% ±  0.6%
  peak_rss           1.74GB ± 17.1KB    1.74GB … 1.74GB          0 ( 0%)        💩+ 44.3% ±  0.0%
  cpu_cycles         37.0G  ± 39.1M     36.9G  … 37.0G           0 ( 0%)        ⚡- 44.4% ±  0.4%
  instructions       78.7G  ± 10.4K     78.7G  … 78.7G           0 ( 0%)        ⚡- 21.2% ±  0.3%
  cache_references   97.5M  ±  524K     97.2M  … 98.1M           0 ( 0%)        ⚡- 76.7% ±  3.4%
  cache_misses       46.5M  ±  159K     46.3M  … 46.6M           0 ( 0%)        💩+ 25.8% ±  1.1%
  branch_misses       327M  ±  116K      327M  …  327M           0 ( 0%)        💩+158.4% ±  0.6%
```

```
Benchmark 1 (3 runs): gcc-13 -fsyntax-only -I/home/ehaas/source/zig/lib /home/ehaas/source/zig/build/zig2.c
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          92.0s  ±  281ms    91.7s  … 92.3s           0 ( 0%)        0%
  peak_rss           1.75GB ±  116KB    1.75GB … 1.75GB          0 ( 0%)        0%
  cpu_cycles          349G  ± 1.21G      348G  …  350G           0 ( 0%)        0%
  instructions       75.0G  ±  103M     75.0G  … 75.2G           0 ( 0%)        0%
  cache_references   6.65G  ± 14.8M     6.64G  … 6.67G           0 ( 0%)        0%
  cache_misses       3.87G  ± 35.7M     3.84G  … 3.91G           0 ( 0%)        0%
  branch_misses       129M  ±  185K      129M  …  130M           0 ( 0%)        0%
Benchmark 2 (3 runs): /home/ehaas/source/arocc/zig-out/bin/arocc-node-locs -fsyntax-only -I/usr/lib/gcc/x86_64-linux-gnu/10/include -I/home/ehaas/source/zig/lib /home/ehaas/source/zig/build/zig2.c
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          10.7s  ±  250ms    10.4s  … 10.9s           0 ( 0%)        ⚡- 88.4% ±  0.7%
  peak_rss           1.74GB ± 49.5KB    1.74GB … 1.74GB          0 ( 0%)          -  0.7% ±  0.0%
  cpu_cycles         37.1G  ± 83.0M     37.0G  … 37.2G           0 ( 0%)        ⚡- 89.4% ±  0.6%
  instructions       78.7G  ±  413      78.7G  … 78.7G           0 ( 0%)        💩+  4.9% ±  0.2%
  cache_references   97.6M  ± 2.19M     95.4M  … 99.8M           0 ( 0%)        ⚡- 98.5% ±  0.4%
  cache_misses       46.1M  ±  152K     45.9M  … 46.2M           0 ( 0%)        ⚡- 98.8% ±  1.5%
  branch_misses       327M  ±  104K      326M  …  327M           0 ( 0%)        💩+152.5% ±  0.3%
```